Incorporate jarmonik's PR (extern_mfd) from test branch.

+++ Src/Orbiter/Planet.h
@@ -183,7 +183,7 @@ public:
 	TileManager2<CloudTile> *CloudMgr2() const { return cmgr2; }
 
 	intptr_t FindFirst (int type, _finddata_t *fdata, char *path, char *fname);
-	int FindNext (intptr_t fh, _finddata_t *fdata, char *fname);
+	intptr_t FindNext (intptr_t fh, _finddata_t *fdata, char *fname);
 
 	//USERLABELSPEC const *GetUserLabel (int idx) const { return userlabel+idx; }
 	//int nUserLabel () const { return nuserlabel; }
